Add a DatabaseManager class
This class will hold the database functions of the history plugin.
